rename spring.adapter to adapter.spring package
