+++ java-security/src/main/java/com/sap/cloud/security/adapter/spring/SAPOfflineTokenServicesCloud.java
@@ -1,4 +1,4 @@
-package com.sap.cloud.security.spring.adapter;
+package com.sap.cloud.security.adapter.spring;
 
 import com.sap.cloud.security.config.OAuth2ServiceConfiguration;
 import com.sap.cloud.security.token.Token;
